+++ src/main/java/io/red/personapi/controllers/requests/AddressDataRequest.java
@@ -1,0 +1,9 @@
+package io.red.personapi.controllers.requests;
+
+import java.util.List;
+
+public record AddressDataRequest(
+        Long personId,
+        List<AddressRequest> addressList
+) {
+}