+++ src/main/java/io/red/personapi/controllers/requests/AddressRequest.java
@@ -1,0 +1,12 @@
+package io.red.personapi.controllers.requests;
+
+public record AddressRequest(
+        String street,
+        String postalCode,
+        String number,
+        String city,
+        boolean mainAddress,
+        Long personId
+) {
+
+}