+++ src/main/java/io/red/personapi/controllers/responses/AddressResponse.java
@@ -1,0 +1,12 @@
+package io.red.personapi.controllers.responses;
+
+public record AddressResponse(
+        Long personID,
+        String street,
+        String postalCode,
+        String number,
+        String city,
+        boolean mainAdress
+
+) {
+}