+++ CHANGELOG.md
@@ -7,16 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 * Search a person;
-* List person's address;
 * Set main person's main address;
 
-## [0.4.0]() - 2020-05-15
+## [0.5.0]() - 2023-05-16 
+### Added
+* Create address to a person;
+* List person's address;
+
+## [0.4.0]() - 2023-05-15
 ### Added
 * List persons;
 * List persons paginated;
 * List persons by ID;
 
-## [0.3.0]() - 2020-05-15
+## [0.3.0]() - 2023-05-15
 ### Added
 * Update a person;